refactor(file-location): move hooks from lib to model segment
