+++ src/features/booking/lib/map-prisma-error.ts
@@ -1,7 +1,7 @@
 import { Prisma } from '@/prisma/generated/client';
 
-import { type ErrorResponse } from '../../api/create-booking';
-import { type UniqueField } from '../../model/booking-form.types';
+import { type ErrorResponse } from '../api/create-booking';
+import { type UniqueField } from '../model/booking-form.types';
 
 const uniqueFieldMessages: Record<UniqueField, string> = {
   phone: 'Пользователь с таким телефоном уже оставил заявку',
